pyproject.toml
[project]
name = "wizard"
version = "0.1.0"
description = "The greatest project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "argcomplete==3.6.3",
    "attrs==25.4.0",
    "blinker==1.9.0",
    "Brotli==1.1.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "curl_cffi==0.7.3",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "Flask==3.1.0",
    "frozenlist==1.8.0",
    "greenlet==3.3.0",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "h2==4.1.0",
    "hpack==4.0.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "hyperframe==6.0.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.4",
    "jiter==0.12.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.3",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.7",
    "langchain-openai==1.1.7",
    "langchain-text-splitters==1.1.0",
    "langgraph==1.0.6",
    "langgraph-checkpoint==4.0.0",
    "langgraph-prebuilt==1.0.6",
    "langgraph-sdk==0.3.3",
    "langsmith==0.6.2",
    "MarkupSafe==3.0.2",
    "marshmallow==3.26.2",
    "msgpack==1.1.0",
    "multidict==6.7.0",
    "mypy_extensions==1.1.0",
    "numpy==2.4.1",
    "openai==2.15.0",
    "orjson==3.11.5",
    "ormsgpack==1.12.1",
    "packaging==24.2",
    "peewee==3.17.8",
    "pip==24.2",
    "pipx==1.8.0",
    "platformdirs==4.5.1",
    "propcache==0.4.1",
    "pycparser==2.22",
    "pydantic==2.12.5",
    "pydantic_core==2.41.5",
    "pydantic-settings==2.12.0",
    "PyExecJS==1.5.1",
    "PyMySQL==1.1.1",
    "pyproject-freeze==0.1.1",
    "PySocks==1.7.1",
    "python-dotenv==1.2.1",
    "PyYAML==6.0.2",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "setuptools==75.6.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.45",
    "tavily-python==0.7.17",
    "tenacity==9.1.2",
    "tiktoken==0.12.0",
    "tomlkit==0.13.2",
    "tqdm==4.67.1",
    "typing_extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.2.3",
    "userpath==1.9.2",
    "uuid_utils==0.13.0",
    "Werkzeug==3.1.3",
    "wheel==0.45.1",
    "win-inet-pton==1.1.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.23.0",
]
